fix: extract validate pool liquidity config
